Copying and deleting folders to destination recursively

+++ copyLib.js
@@ -13,29 +13,36 @@ copy.removeDirectoryRecursively = function(directory){
 	if(!fs.existsSync(directory)) return;
 
 	var files = copy.filesInDirectory(directory);
+	var path;
+
 	files.forEach(function(file){
+		path = directory + '/' + file;
+		if(fs.lstatSync(path).isDirectory()) return copy.removeDirectoryRecursively(path);
 		fs.unlinkSync(directory + '/' + file);
 	});
 	fs.rmdirSync(directory);
 };
 
-copy.copyFileToDestination = function(sourse , destination , file){
+copy.copyFileToDestination = function(source , destination , file){
 	var writeLocation = destination +'/'+ file;
+	var readLocation = source + '/' + file;
+
+	if(fs.lstatSync(readLocation).isDirectory()) return copy.copyFiles(readLocation , writeLocation);
 	if(fs.existsSync(writeLocation)) return ;
 	
-	var text = fs.readFileSync(sourse + '/' + file);
+	var text = fs.readFileSync(readLocation);
 	console.log("\n.........Copying "+ file + " to " + destination)
 	fs.writeFileSync(writeLocation, text);
 };
 
-copy.copyFiles = function(sourse , destination , option){
+copy.copyFiles = function(source , destination , option){
 
 	if(option == '--all') copy.removeDirectoryRecursively(destination);
 	copy.createDir(destination);
 
-	var sourseFiles = copy.filesInDirectory(sourse);
-	sourseFiles.forEach(function(file){
-		copy.copyFileToDestination(sourse , destination , file);
+	var sourceFiles = copy.filesInDirectory(source);
+	sourceFiles.forEach(function(file){
+		copy.copyFileToDestination(source , destination , file);
 	})
 };
